separated json serde code into multiple class files
