+++ src/main/java/de/azapps/kafkabackup/common/record/ConnectHeaderJSONSerde.java
@@ -1,0 +1,6 @@
+package de.azapps.kafkabackup.common.record;
+
+public class ConnectHeaderJSONSerde {
+    protected static final String HEADER_KEY_PROPERTY = "key";
+    protected static final String HEADER_VALUE_PROPERTY = "value";
+}
